Refactor Yjs sync: Remove registry, simplify system management
