+++ js/summarization.js
@@ -42,13 +42,8 @@ const saveSummaries = (summaries) => {
   const result = safeSetToStorage(STORAGE_KEY, summaries);
   
   // Update Yjs directly if available (for real-time sync)
-  import('./yjs-registry.js').then(({ getYjsSystem }) => {
-    const yjsSystem = getYjsSystem();
-    if (yjsSystem) {
-      import('./yjs.js').then(({ updateSummariesInYjs }) => {
-        updateSummariesInYjs(yjsSystem, summaries);
-      });
-    }
+  import('./yjs.js').then(({ updateSummariesInYjs }) => {
+    updateSummariesInYjs(summaries);
   }).catch(() => {
     // Yjs not available, that's OK
   });
